pyproject.toml
[tool.poetry]
name = "memorylane"
version = "0.1.0"
description = ""
authors = ["Deeptanshu Sanghani <deeptanshusanghani@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.114.0"
uvicorn = {extras = ["standard"], version = "^0.30.6"}
python-multipart = "^0.0.9"
google-cloud-storage = "^2.18.2"
sqlalchemy = "^2.0.34"
pydantic = "^2.9.0"
strawberry-graphql = {extras = ["fastapi"], version = "^0.240.0"}
packaging = "^24.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
